refactoring v1
