--- MuTe_RL/models/rubiks/__init__.py
@@ -1,3 +1,0 @@
-from MuTe_RL.models.rubiks.look_ahead import score_actions, search_best_paths
-
-__all__ = ["score_actions", "search_best_paths"]
