--- MuTe_RL/simulations/rubiks/__init__.py
@@ -1,6 +1,0 @@
-from .cube import Cube
-from .layer import Layer
-from .rubik import Rubik
-from .color import Colors, BackgroundColors
-
-__all__ = ["Cube", "Layer", "Rubik", "Colors", "BackgroundColors"]
