--- MuTe_RL/simulations/rubiks/rubik.py
@@ -1,9 +1,0 @@
-"""simulate any rubik's cube or any other polyhedron"""
-
-from abc import ABC
-
-
-
-class Rubik(ABC):
-    pass
-
