+++ MuTe_Rubik/models/rubiks/__init__.py
@@ -1,0 +1,3 @@
+from MuTe_Rubik.models.rubiks.look_ahead import score_actions, search_best_paths
+
+__all__ = ["score_actions", "search_best_paths"]
