+++ MuTe_Rubik/models/rubiks/look_ahead.py
