+++ requirements.txt
@@ -1,4 +1,3 @@
 torch
 numpy
-ipykernel
 tqdm
